refactored the structure and names of directories/repositories/classes to make the code more readable
