+++ src/main/java/com/liftoff/controllers/admin/AdminWantToController.java
@@ -3,7 +3,7 @@ package com.liftoff.controllers.admin;
 import com.liftoff.models.WantTo;
 import com.liftoff.models.Welcome;
 import com.liftoff.models.data.WantToRepository;
-import com.liftoff.models.data.WelcomeRepository;
+import com.liftoff.models.data.UIRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -19,7 +19,7 @@ import java.util.List;
 public class AdminWantToController {
 
     @Autowired
-    private WelcomeRepository welcomeRepository;
+    private UIRepository UIRepository;
 
     @Autowired
     private WantToRepository wantToRepository;
@@ -27,7 +27,7 @@ public class AdminWantToController {
     @GetMapping ("")
     public String displayManageWantTosPage (Model model) {
         List<WantTo> listWantTo = wantToRepository.findAll();
-        Welcome welcome = welcomeRepository.getById(1);
+        Welcome welcome = UIRepository.getById(1);
         model.addAttribute("title", "Admin Portal: Manage Want-To");
         model.addAttribute("listWantTo", listWantTo);
         model.addAttribute("welcome", welcome);
