+++ src/main/java/com/liftoff/controllers/site/HelpController.java
@@ -3,7 +3,7 @@ package com.liftoff.controllers.site;
 
 import com.liftoff.models.data.FaqRepository;
 import com.liftoff.models.Faq;
-import com.liftoff.models.data.WelcomeRepository;
+import com.liftoff.models.data.UIRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import org.springframework.stereotype.Controller;
@@ -22,7 +22,7 @@ public class HelpController {
     private FaqRepository faqRepository;
 
     @Autowired
-    private WelcomeRepository welcomeRepository;
+    private UIRepository UIRepository;
