+++ src/main/java/com/liftoff/models/data/UIRepository.java
@@ -3,5 +3,5 @@ package com.liftoff.models.data;
 import com.liftoff.models.Welcome;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WelcomeRepository extends JpaRepository<Welcome, Integer> {
+public interface UIRepository extends JpaRepository<Welcome, Integer> {
 }
