+++ src/main/resources/templates/account/index.html
@@ -16,6 +16,10 @@
                         <div class="text-center p-3">
                             <button type="submit" class="btn btn-success">logout</button>
                         </div>
+<!--                        <div>-->
+<!--                            <p th:text="${user.username}"> </p>-->
+<!--                            <p th:text="${user.password}"> </p>-->
+<!--                        </div>-->
                     </div>
                 </form>
             </div>
