+++ src/main/resources/templates/fragments.html
@@ -145,7 +145,7 @@
     <nav th:fragment="adminNav" class= "navbar bg-light" sec:authorize="hasAnyAuthority('ADMIN','EDITOR')" >
         <ul class="navbar-nav">
             <li class="nav-item"><a class="nav-link" th:href="@{/account/manageUsers}">Users</a></li>
-            <li class="nav-item"><a class="nav-link" th:href="@{/account/manageWelcome/edit}">UI</a></li>
+            <li class="nav-item"><a class="nav-link" th:href="@{/account/manageUI/edit}">UI</a></li>
             <li class="nav-item"><a class="nav-link" th:href="@{/account/manageFaqs}">FAQs</a></li>
             <li class="nav-item"><a class="nav-link" th:href="@{/account/manageWantTos}">WantTo(s)</a></li>
             <li class="nav-item"><a class="nav-link" th:href="@{/account/manageLinks}">Links</a></li>
