--- src/main/resources/templates/users/new.html
@@ -1,57 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org/">
-    <head th:replace="fragments :: head"></head>
-
-    <body>
-
-    <div class="container text-center">
-
-        <header th:replace ="fragments :: header"></header>
-
-        <form th:action="@{/users/save}" th:object="${user}" th:method="post"
-                style="max-width: 600px; margin: 0; auto">
-            <input type="hidden" th:field="*{id}" th:value="${user.id}"/>
-            <div class="m-3">
-
-                <div Class="form-group row">
-                    <label class="col-form-Label col-sm-4">username:</label>
-                    <div class="col-sm-8">
-                        <input type="text" th:field="*{username}" class="form-control" required/>
-                    </div>
-                </div>
-
-                <div Class="form-group row">
-                    <label class="col-form-Label col-sm-4">User email:</label>
-                    <div class="col-sm-8">
-                        <input type="text" th:field="*{email}" class="form-control"/>
-                    </div>
-                </div>
-
-                <div Class="form-group row">
-                    <label class="col-form-Label col-sm-4">User password:</label>
-                    <div class="col-sm-8">
-                        <input type="text" th:field="*{pwHash}" class="form-control" required/>
-                    </div>
-                </div>
-
-                <div Class="form-group row">
-                    <label class="col-form-Label col-sm-4">User roles:</label>
-                    <div class="col-sm-8 text-left">
-                        <th block th:each="role : ${listRoles}" >
-                            <div>
-                                <input type="checkbox" th:field="*{roles}"
-                                       th:text="${role.name}"
-                                       th:value="${role.id}" />
-                            </div>
-                        </th>
-                    </div>
-                </div>
-
-                <div class="text-center p-3">
-                    <button type="submit" class="btn btn-primary">Add</button>
-                </div>
-            </div>
-        </form>
-    </div>
-    </body>
-</html>
